refactor(healthplatform): rename admisconfig package directory to ad-misconfiguration
Renames comp/healthplatform/issues/admisconfig/ to
comp/healthplatform/issues/ad-misconfiguration/ so the directory name
aligns with the kebab-case convention used for issue ID prefixes
(AnnotationIssueID = "ad-annotation", TemplateIssueID = "ad-template").

The Go package name remains admisconfig throughout — only the import
path and directory name change. Updates all import paths, BUILD deps,
and the Bazel importpath declaration.
