+++ comp/core/autodiscovery/impl/BUILD.bazel
@@ -40,7 +40,7 @@ go_library(
         "//comp/core/workloadfilter/def",
         "//comp/core/workloadmeta/def",
         "//comp/def",
-        "//comp/healthplatform/issues/admisconfig",
+        "//comp/healthplatform/issues/ad-misconfiguration",
         "//comp/healthplatform/store/def",
         "//pkg/collector/check",
         "//pkg/collector/check/id",
@@ -94,8 +94,7 @@ go_test(
         "//comp/core/workloadmeta/def",
         "//comp/core/workloadmeta/fx-mock",
         "//comp/core/workloadmeta/mock",
-        "//comp/healthplatform/issues/admisconfig",
-        "//comp/healthplatform/store/def",
+        "//comp/healthplatform/issues/ad-misconfiguration",
         "//comp/healthplatform/store/mock",
         "//comp/healthplatform/store/noop-impl",
         "//pkg/collector/check/id",
