+++ comp/healthplatform/bundle.go
@@ -26,12 +26,12 @@ import (
 	// Issue modules register themselves via init(); imported here for side effects.
 	registrydef "github.com/DataDog/datadog-agent/comp/healthplatform/issueregistry/def"
 	registryfx "github.com/DataDog/datadog-agent/comp/healthplatform/issueregistry/fx"
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/admisconfig"       // registers templates via init()
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/admissionprobe"    // registers templates via init()
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/checkfailure"      // registers templates via init()
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/dockerpermissions" // registers templates via init()
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/invalidconfig"     // registers templates via init()
-	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/rofspermissions"   // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/ad-misconfiguration" // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/admissionprobe"      // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/checkfailure"        // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/dockerpermissions"   // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/invalidconfig"       // registers templates via init()
+	_ "github.com/DataDog/datadog-agent/comp/healthplatform/issues/rofspermissions"     // registers templates via init()
 	runnerdef "github.com/DataDog/datadog-agent/comp/healthplatform/runner/def"
 	runnerfx "github.com/DataDog/datadog-agent/comp/healthplatform/runner/fx"
 	schedulerdef "github.com/DataDog/datadog-agent/comp/healthplatform/scheduler/def"
