+++ comp/healthplatform/store/def/BUILD.bazel
@@ -2,10 +2,7 @@ load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "def",
-    srcs = [
-        "component.go",
-        "constants.go",
-    ],
+    srcs = ["component.go"],
     importpath = "github.com/DataDog/datadog-agent/comp/healthplatform/store/def",
     visibility = ["//visibility:public"],
     deps = ["@com_github_datadog_agent_payload_v5//healthplatform"],
